Rev5 checkpoint: Routed most of Serial Flash
All of the routing difficulties for the two Serial
EEPROM chips have been solved now.

-The perilous routes to the GPIO pins from the AND
 gate appear to fit!

-The shared data and clock lines on the Serial
 EEPROM chips have been linked using vias.

There are still some tough problems to solve.
There is not going to be quite enough room to
squeeze the last signals into the AND gate chip
as-is, and routing the control lines from the
edge connector in general is looking tricky.