First chunk of work on new logo/back silkscreen
Figured out and documented (in skunk_v5.txt) a
workflow for creating a gerber file from an
Inkscape SVG file. Currently, back_silk.svg is a
near-exact duplicate of the original back silk
screen, but now that I've worked out the process,
it will be trivial to make the minor updates I had
in mind.