Rev. 5 RC2 - Move U6 to fit in standard cart case
I realized on Rev. 4 and Rev. 5 RC1 boards that U6
is in the way of a support structure in the Jaguar
cartridge cases. Moving U6 over rather close to
X1 and rerouting some traces around it (needed one
more pair of vias) should get it *just* far enough
out of the way to fit.

The lowest serial EEPROM trace routed through the
finger portion of the PCB was also raised up a bit
to better ensure it clears the edge bevel. It
hadn't caused any problems at the lower location
yet, but it was cutting it closer than I'd thought
it would. It now overlaps another trace in the
neighboring inner layer. Hopefully this doesn't
result in noticeable interference.

Also added support for solder-paste/stencil gerber
files to the release/packaging infrastructure in
mkfab.sh.