Fix svg2shenzen version of back silkscreen gerber
For some reason, KiCad insists on putting a thing
outline around the filled regions in its
components, which makes its way into the gerber
file. It's thin enough that it can either round
down to a degenerate zero-width object or a
1-unite-wide line, likely depending on the
software used to view/apply the gerber file and/or
the options set in it. I had it come back realized
as a zero-width line in one build, and a 1-width
line in another, both from the same manufacturer.

To fix this, I manually removed the outlines
entirely from the gerber file. This involved
deleting all the pen movement commands between
G37 and G36 sections (begin and end filled region
respectively).